==> 1 - IMPORTANTE Preparação base de dados.ipynb <==
# Considerações iniciais
Essa base de dados foi retirada de uma competição do kaggle, porém foi modificada.
# A competição original era sobre previsão de vendas de mais de 1000 lojas, por limitações de hardware e utilização de portfólio eu reduzi para 20 lojas.
#### Link da competição: www.kaggle.com/c/rossmann-store-sales/

# Importação das bibliotecas

import pandas as pd
import numpy as np
from sklearn.impute import SimpleImputer
from sklearn.metrics import mean_squared_error
from lightgbm import LGBMRegressor
pd.options.display.max_columns = None

total_lojas=20

# Leitura da base de dados

train = pd.read_csv('train.csv')

# Limitação a vinte lojas

Conforme dito, irei reduzir a 20 lojas por questões de performance.

train = train[train['Store'] <= total_lojas].reset_index(drop=True)

train.head(25)

# Pré processamento

Na coluna StateHoliday há um problema de haver 0 em inteiro e 0 em string. Portanto, precisamos ajustar esse erro.

def trocar(x):
    #Poderia ser lambda
    if x ==0:
        return '0'
    else:
        return x
    
train['StateHoliday'] = train['StateHoliday'].map(trocar)
train['StateHoliday'].value_counts()

Após fazer a transformação nos dados, optei por criar uma coluna ao estilo OneHotEncoder sinalizando a informação.

train['StateHoliday'] = 'StateHoliday_'+ train['StateHoliday']
train = pd.concat([train.drop(columns='StateHoliday'),pd.get_dummies(train['StateHoliday'])],axis=1)
train

Para adicionar novas informações ao dataset, utilizarei o outro csv disponível e os juntarei. Novamente limitando o total de lojas.

store = pd.read_csv('store.csv')
store = store[store['Store'] <=total_lojas].reset_index(drop=True)
store.head(5)

## Preenchimento dos valores nulos
Eu separei difentes classes para assim eu possa usar estratégias diferentes para valores numéricos e categóricos. Para esse caso, utilizei os valores mais frequentes.

float_columns=['CompetitionOpenSinceMonth','CompetitionOpenSinceYear','Promo2SinceWeek','Promo2SinceYear']
float_inputer = SimpleImputer(strategy='most_frequent',add_indicator=False)
float_inputer.fit(store[float_columns])

categorical_columns=['PromoInterval']
categorical_inputer = SimpleImputer(strategy='most_frequent',add_indicator=False)
categorical_inputer.fit(store[categorical_columns])

store[float_columns] = float_inputer.transform(store[float_columns])
store[categorical_columns] = categorical_inputer.transform(store[categorical_columns])
store.head(5)

## Transformações nas colunas StoreType e Assortment

Se utilizando novamente da estratégia para transformar valores em colunas, eu utilizo para as colunas StoreType e Assortment.

store = pd.concat([store.drop(columns=['StoreType','Assortment']),pd.get_dummies(store[['StoreType','Assortment']])],axis=1)
store.head(5)

Restando apenas a coluna PromoInterval para processar, utilizo uma função simples para transformar os valores em numéricos.

def trocar2(x):
    if x =='Jan,Apr,Jul,Oct':
        return 1
    elif x == 'Mar,Jun,Sept,Dec':
        return 2 
    else:
        return 3
store['PromoInterval'] = store['PromoInterval'].map(trocar2)
store.head(3)

# Juntando os dois datasets através de um merge

train = train.merge(store,how='left',on='Store')

train['Date'] = pd.to_datetime(train['Date'])
train = train.sort_values(by=['Date','Store']).reset_index(drop=True)

Como o dataset está ordenado inversamente, por questão de costume eu troco para valores mais antigos primeiros, seguido pelo numero da loja.

## Criando uma coluna binária referente a cada loja 

train['Store'] = ('Store_'+train['Store'].astype(str))
aux = pd.get_dummies(train['Store'])
train = pd.concat([train.drop(columns='Store'),aux],axis=1)

Conforme dito, há 20 lojas. Eu optei por criar uma coluna binaria para cada loja.
Como método eu transformei os valores respectivos a cada loja em valor de string.

Exemplo: 5 --> '5'

Então, adicionei um texto informando o nome, passando a ser:

'5' --> 'Store_5'
 
E finalmente, criando uma coluna para cada loja.

train

def datas(df):
    df['Date'] = pd.to_datetime(df['Date'])
    
    df['sin_dia_do_mes'] = np.sin(2*np.pi*(df['Date'].dt.day)/31)
    df['cos_dia_do_mes'] = np.cos(2*np.pi*(df['Date'].dt.day)/31)

    df['sin_mes'] = np.sin(2*np.pi*(df['Date'].dt.month)/12)
    df['cos_mes'] = np.cos(2*np.pi*(df['Date'].dt.month)/12)

    return df
train = datas(train)
train.head(5)

Para a representação das datas, além do dia da semana já informado no dataset, eu utilizo a técnica de fazer o seno e cosseno das datas, na esperança que os modelos captem suas ciclicidades.

## Separando em Train_lvl0, Train_lvl1 e Valid

Para este projeto, utilizarei um holdout simples separando o train_lvl1 e valid com 60 dias.

train['Date'].describe()

train_lvl0 = train[(train['Date'] <= '2015-04-02')]

train_lvl1 = train[(train['Date'] > '2015-04-02')]
train_lvl1 = train_lvl1[train_lvl1['Date'] <= '2015-06-01']

valid = train[train['Date'] > '2015-06-01']

print(train_lvl0['Date'].describe())
print(train_lvl1['Date'].describe())
print(valid['Date'].describe())

# Definindo a baseline

Definir a baseline é um passo muito importante, pois com elas conseguimos ter uma noção sobre como nossos modelos estão se saindo. Neste projeto, irei utilizar as vendas do dia anterior como baseline.

baseline = pd.DataFrame()
baseline['Date'] = train['Date']
baseline['Sales_original'] = train['Sales']
baseline['Lag_one_day'] = train['Sales'].shift(20).fillna(0)

aux = baseline[baseline['Date'] <= '2015-04-02'] 
from sklearn.metrics import mean_squared_error
print('Baseline train 0')
print('RMSE:',np.sqrt(mean_squared_error(aux['Sales_original'],aux['Lag_one_day'])))
print('RMSLE:',np.log(np.sqrt(mean_squared_error(aux['Sales_original'],aux['Lag_one_day']))))
print('Mae:',mean_squared_error(aux['Sales_original'],aux['Lag_one_day']))

aux = baseline[baseline['Date'] > '2015-04-02'] 
aux = aux[aux['Date'] <= '2015-06-01']
from sklearn.metrics import mean_squared_error
print('Baseline train 1')
print('RMSE:',np.sqrt(mean_squared_error(aux['Sales_original'],aux['Lag_one_day'])))
print('RMSLE:',np.log(np.sqrt(mean_squared_error(aux['Sales_original'],aux['Lag_one_day']))))
print('Mae:',mean_squared_error(aux['Sales_original'],aux['Lag_one_day']))

aux = baseline[baseline['Date'] > '2015-06-01'] 
from sklearn.metrics import mean_squared_error
print('Baseline valid')
print('RMSE:',np.sqrt(mean_squared_error(aux['Sales_original'],aux['Lag_one_day'])))
print('RMSLE:',np.log(np.sqrt(mean_squared_error(aux['Sales_original'],aux['Lag_one_day']))))
print('Mae:',mean_squared_error(aux['Sales_original'],aux['Lag_one_day']))

Iremos utilizar o RMSE como nossa baseline.

# Pré-processamento na base de teste
Apenas repetindo os passos feitos com a base de treino. A base de teste é fornecida na competição.

test = pd.read_csv('test.csv')

test = test[test['Store'] <= total_lojas].reset_index(drop=True)
test['StateHoliday'] = test['StateHoliday'].map(trocar)
test['StateHoliday'] = 'StateHoliday_'+ test['StateHoliday']
test = pd.concat([test.drop(columns='StateHoliday'),pd.get_dummies(test['StateHoliday'])],axis=1)
test = test.merge(store,how='left',on='Store')
test['Date'] = pd.to_datetime(test['Date'])
test = datas(test)
test['Store'] = ('Store_'+test['Store'].astype(str))
aux = pd.get_dummies(test['Store'])
test = pd.concat([test.drop(columns='Store'),aux],axis=1)

test

## Garantindo que base de teste tem as mesmas colunas que a base de treino

test2 = pd.DataFrame()
test2['Id'] = test['Id']
missing_cols = set( train.columns ) - set( test.columns )

for c in missing_cols:
    test[c] = 0

test = test[train.columns]
#Coluna a mais - sale
#Coluna removida - ID
test = pd.concat([test2,test.drop(columns='Sales')],axis=1)

## Salvando as bases

train_lvl0.to_csv('train_lvl0.csv',index=False)
train_lvl1.to_csv('train_lvl1.csv',index=False)
valid.to_csv('valid.csv',index=False)
test.to_csv('teste.csv',index=False)
