==> 3 - Seleção de modelos.ipynb <==
# Importações

import pandas as pd
import numpy as np
import joblib as jb
import glob
import re
from sklearn.linear_model import Ridge
from sklearn.metrics import mean_squared_error

# Leitura e separação das bases em X e Y

train_lvl0 = pd.read_csv('train_lvl0.csv')
train_lvl1 = pd.read_csv('train_lvl1.csv')
val = pd.read_csv("valid.csv")

X_train0, Y_train0 = train_lvl0.drop(columns=['Date','Sales']), train_lvl0['Sales']
X_train1, Y_train1 = train_lvl1.drop(columns=['Date','Sales']), train_lvl1['Sales']
X_val, Y_val = val.drop(columns=['Date','Sales']), val['Sales']

# Carregamento de todas as previsões feitas 
Conforme dito no notebook anterior, para gerar o stacking precisamos de diversidades. Então aqui iremos carregar todas as previsões feitas pelos modelos para o train1 e val e transformar em um dataframe de previsões

preds_train1 = glob.glob("./preds_train1/*.pkl.z")
preds_val = glob.glob("./preds_val/*.pkl.z")

df_train1 = []
for p_name in preds_train1:
    p = jb.load(p_name)
    p_name = re.search(r"train1\\(.*)\.pkl.z", p_name).group(1)
    p_df = pd.DataFrame(p, columns=[p_name])
    df_train1.append(p_df)
    
    
df_val = [] 
for p_name in preds_val:
    p = jb.load(p_name)
    p_name = re.search(r"val\\(.*)\.pkl.z", p_name).group(1)
    p_df = pd.DataFrame(p, columns=[p_name])
    df_val.append(p_df)
    
df_train1 = pd.concat(df_train1, axis=1)
df_val = pd.concat(df_val, axis=1)

Por mais que pareça um código complexo, apenas pense que as previsões feitas anteriormente foram carregadas e salvas em um Dataframe.

df_train1

# Seleção das melhores combinações de modelos
Uma vez carregados, precisamos avaliar as previsões e encontrar quais delas combinadas geram o melhor resultado. Começamos sem selecionar nenhuma, pois não sabemos qual a melhor. 

Para esse projeto, utilizarei uma ridge regression por ser um modelo que performa relativamente rápido, para cada previsão gerada no passo 2, a ridge irá tentar utilizar como feature, e naturalmente as que conseguirem os melhores resultados serão exibidas.

Em suma: estamos procurando as colunas, ou seja previsões feitas no notebook anterior, que nos dê o melhor resultado.

selected = [
'lgbm_110_14_0.03248168667854209_215',
'lgbm_361_59_0.05378358562195618_276',
'rf_176_3_48',
'lgbm_24_1_0.03199227519485914_200',
'lgbm_185_96_0.014121042963223253_287',
'svm_rbf_0.6614584754426877',
'svm_rbf_8.713863857748523',
'lgbm_238_80_0.05209570020716538_268',
'lgbm_427_1_0.0001_236',
'rf_139_17_34',
'lgbm_55_48_0.018714601098343324_274']

best_score = 1500
for col in df_train1.columns:
    if col not in selected:
        Xtr = df_train1[selected+[col]].copy()
        Xval = df_val[selected+[col]].copy()
    
        mdl = Ridge(alpha=1.)
        mdl.fit(Xtr, Y_train1)

        p = mdl.predict(Xval)
        c = np.sqrt(mean_squared_error(Y_val, p))
        
        if c < best_score:
            print(col, c)
            best_score = c
 

Perceba que ainda não foi otimizado 100%, porém o ganho estava baixo, decidi parar. Dependendo do contexto pode-se extrair até o limite.
