==> 4 - Final.ipynb <==
# Considerações iniciais
Conforme feito nos notebooks anterios, nós preparamos os dados, geramos diversidades, e selecionamos as melhores combinações de modelos. Agora iremos apenas ajustar os modelos e persistir na memória.

# Importações

import pandas as pd 
import numpy as np
from lightgbm import LGBMRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.svm import SVR
from sklearn.metrics import mean_squared_error
import pickle as pkl

# Leitura e separação das bases em X e Y

train_lvl0 = pd.read_csv('train_lvl0.csv')
train_lvl1 = pd.read_csv('train_lvl1.csv')
val = pd.read_csv("valid.csv")

X_train0, Y_train0 = train_lvl0.drop(columns=['Date','Sales']), train_lvl0['Sales']
X_train1, Y_train1 = train_lvl1.drop(columns=['Date','Sales']), train_lvl1['Sales']
X_val, Y_val = val.drop(columns=['Date','Sales']), val['Sales']

preds_lvl1 = pd.DataFrame()
preds_val = pd.DataFrame()

# Criando uma lista com os modelos
Copiando os parâmetros dos modelos selecionados no notebook anterior, criamos uma lista com cada modelo, para assim facilitar na implementação.

models = [
        LGBMRegressor(num_leaves=110,
                        min_data_in_leaf=14,
                        learning_rate=0.03248168667854209, 
                        n_estimators=215,
                        random_state=0),
    
        LGBMRegressor(num_leaves=361,
                        min_data_in_leaf=59,
                        learning_rate=0.05378358562195618, 
                        n_estimators=276,
                        random_state=0),

        RandomForestRegressor(n_estimators=176,
                                       max_depth=3,
                                       min_samples_leaf=48,
                                       n_jobs=-1),
    
        LGBMRegressor(num_leaves=24,
                        min_data_in_leaf=1,
                        learning_rate=0.03199227519485914, 
                        n_estimators=200,
                        random_state=0),
    
        LGBMRegressor(num_leaves=185,
                        min_data_in_leaf=96,
                        learning_rate=0.014121042963223253, 
                        n_estimators=287,
                        random_state=0),

        SVR(kernel='rbf',
        C=0.6614584754426877),

        SVR(kernel='rbf',
        C=8.713863857748523),
    
        LGBMRegressor(num_leaves=238,
                        min_data_in_leaf=80,
                        learning_rate=0.05209570020716538, 
                        n_estimators=268,
                        random_state=0),
    
        LGBMRegressor(num_leaves=427,
                        min_data_in_leaf=1,
                        learning_rate=0.0001, 
                        n_estimators=236,
                        random_state=0),
    
        RandomForestRegressor(n_estimators=139,
                                   max_depth=17,
                                   min_samples_leaf=34,
                                   n_jobs=-1),
        LGBMRegressor(num_leaves=55,
                    min_data_in_leaf=48,
                    learning_rate=0.018714601098343324, 
                    n_estimators=274,
                    random_state=0),
    
      
]

# Treinamento dos modelos 
Utilizando umá estrutura de repetição, treinamos cada modelo com base em sua posição na lista.

for i in range(len(models)):
    models[i].fit(X_train0, Y_train0)

# Criando as previsões de primeiro nivel e da validação 
Novamente, percorrendo com uma estrutura de repetição, com base na posição de cada modelo na lista, faço a previsões de cada modelo, e os guardo em dataframes.

for i in range(len(models)):
    prev = models[i].predict(X_train1)
    preds_lvl1['Modelo_{}'.format(i)] = prev

    prev = models[i].predict(X_val)
    preds_val['Modelo_{}'.format(i)] = prev

# Treinando a ridge de 2º nível

Conforme feito no notebook anterior, utilizamos uma ridge no nosso segundo nível, que irá utilizar as previsões dos modelos acima como features.

from sklearn.linear_model import Ridge
ridge = Ridge()
ridge.fit(preds_lvl1,Y_train1)
preds = ridge.predict(preds_val)

np.sqrt(mean_squared_error(Y_val,preds))

# Resultado validação

Conforme vemos, nossa baseline no primeiro nobeook para a validação era cerca de 3866, que eram as vendas utilizadas no dia anterior. E apesar de todo trabalho para chegar até aqui, conseguimos a marca de 385.

# Visualizando as previsões da ridge

preds[0:100]

# Previsões na base de teste

Como essa é uma base retirada e adaptada de uma competição do kaggle, temos uma base de teste, porém os resultados não são 100% aplicáveis, pois reduzimos o número de lojas. Se baseando nesse contexto, conforme feito no notebook 1, faremos a previsão para apenas as 20 primeiras lojas.

test = pd.read_csv('teste.csv')
Sub = pd.DataFrame()
Sub['Id'] = test['Id']
test = test.drop(columns=['Date','Id'])

preds_test = pd.DataFrame()

for i in range(len(models)):
    prev = models[i].predict(test)
    preds_test['Modelo_{}'.format(i)] = prev

p = ridge.predict(preds_test)
p[p < 0] = 0
Sub['Sales'] = p.round().astype(int)
Sub

# Persistindo na memória os modelos para utilizações em caso de deploy

Essa é uma parte muito importante em caso de deploy, pois precisamos salvar os modelos para que possam ser utilizados no futuro em produção.

file_name = "Modelos.pkl"

open_file = open(file_name, "wb")
pkl.dump([models,ridge], open_file)
open_file.close()

# Leitura dos modelos para certeza de que não houve problemas

open_file = open(file_name, "rb")
models,ridge = pkl.load(open_file)

print(models,ridge)

# Considerações finais 

Estou utilizando esses notebooks como forma de demonstrar minhas habilidades que venho adquirindo ao longo dessa minha trajetória como estudante de engenharia de computação e cientista de dados.

Agradeço a você que tenha acompanhado esse projeto até aqui. Como ideia para futuros projetos pode ser interessante recriar com todas as lojas disponíveis no dataset original e não apenas 20.
